+++ src/main/java/com/nw2/parcel/Dtos/DormListItemDto.java
@@ -1,6 +1,6 @@
 package com.nw2.parcel.Dtos;
 
 public record DormListItemDto(
-        Long dormId,
+        Integer dormId,
         String dormName
 ) {}
