+++ src/main/java/com/nw2/parcel/Dtos/RegisterDto.java
@@ -20,17 +20,16 @@ public record RegisterDto(
         @NotBlank(message = "Role is required")   // "RESIDENT" หรือ "STAFF"
         String role,
 
-        // ✅ ใหม่ — frontend สามารถส่งได้ทั้ง dormId หรือ dormName
-        Long dormId,
+        // ✅ frontend สามารถส่งได้ทั้ง dormId หรือ dormName
+        Integer dormId,        // ⬅ เปลี่ยนจาก Long → Integer
         String dormName,
 
         // ช่องเพิ่มตาม role
-        String roomNumber,   // จำเป็นเมื่อ role=RESIDENT
-        String position      // จำเป็นเมื่อ role=STAFF
+        String roomNumber,     // จำเป็นเมื่อ role=RESIDENT
+        String position        // จำเป็นเมื่อ role=STAFF
 ) {
     // ===== Validation เฉพาะ role =====
 
-    // Resident ต้องมี dormId หรือ dormName อย่างใดอย่างหนึ่ง
     @AssertTrue(message = "Dorm ID or Name is required for RESIDENT")
     @JsonIgnore
     public boolean isDormValidForResident() {
@@ -50,10 +49,10 @@ public record RegisterDto(
         return !isRole("STAFF") || hasText(position);
     }
 
-    // ===== Helper methods =====
     private boolean isRole(String target) {
         return role != null && role.trim().equalsIgnoreCase(target);
     }
+
     private boolean hasText(String s) {
         return s != null && !s.trim().isEmpty();
     }
