+++ src/main/java/com/nw2/parcel/entity/StaffDormId.java
@@ -9,6 +9,7 @@ import java.io.Serializable;
 @NoArgsConstructor
 @AllArgsConstructor
 public class StaffDormId implements Serializable {
-    private Long dormId;
-    private Long userId;
+    private Integer dormId;   // ⬅ เปลี่ยน Long -> Integer
+    private Integer userId;   // ⬅ เปลี่ยน Long -> Integer
 }
+
