+++ src/main/java/com/nw2/parcel/repositories/UsersRepository.java
@@ -4,7 +4,7 @@ import com.nw2.parcel.entity.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UsersRepository extends JpaRepository<Users, Long> {
+public interface UsersRepository extends JpaRepository<Users, Integer> {
     Optional<Users> findByEmail(String email);
     boolean existsByEmail(String email);
     boolean existsByFirebaseUid(String firebaseUid);
